refactor: update flaky tests and project structure
