+++ src/loki.egg-info/PKG-INFO
@@ -1,0 +1,5 @@
+Metadata-Version: 2.4
+Name: loki
+Version: 0.1.0
+License-File: LICENSE
+Dynamic: license-file
