+++ src/loki.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+LICENSE
+README.md
+pyproject.toml
+src/loki/__init__.py
+src/loki/cmd_parser.py
+src/loki/commands.py
+src/loki/main.py
+src/loki.egg-info/PKG-INFO
+src/loki.egg-info/SOURCES.txt
+src/loki.egg-info/dependency_links.txt
+src/loki.egg-info/top_level.txt
+tests/test_cmd_parser.py
+tests/test_commands.py
